Let the log-store resource manage it's own model sub-tree
